[TAS-588]add: 둘러보기 UI - 인증 사진 상세 모달 컨텐츠 좋아요 기능 구현
