+++ src/components/Feed/index.ts
@@ -1,3 +1,4 @@
 export * from './SuccessTaskImageList';
 export * from './SuccessTaskImageCard';
 export * from './SuccessTaskImageDetail';
+export * from './SuccessTaskImageDetailItem';
