+++ src/types/Feed/index.ts
@@ -1,0 +1,10 @@
+// TODO: 인증 사진 리스트 API 연동 시 제거
+type SuccessImageItem = {
+  successImageUrl: string;
+  title: string;
+  profileImageUrl: string;
+  userName: string;
+  likeCount: number;
+};
+
+export type { SuccessImageItem };
